fix getDiskSpace on 32bit systems

+++ adm_program/system/classes/FileSystemUtils.php
@@ -208,7 +208,7 @@ final class FileSystemUtils
      * Gets the total, free and used disk space in bytes
      * @param string $path Path of the filesystem
      * @throws \RuntimeException Throws if the given path is not in an allowed directory
-     * @return array<string,int> Returns the total, free and used disk space in bytes
+     * @return array<string,float> Returns the total, free and used disk space in bytes
      * @see https://secure.php.net/manual/en/function.disk-total-space.php
      * @see https://secure.php.net/manual/en/function.disk-free-space.php
      * @example array("total" => 10737418240, "free" => 2147483648, "used" => 8589934592)
@@ -217,8 +217,8 @@ final class FileSystemUtils
     {
         self::checkIsInAllowedDirectories($path);
 
-        $total = (int) disk_total_space($path);
-        $free = (int) disk_free_space($path);
+        $total = disk_total_space($path);
+        $free = disk_free_space($path);
         $used = $total - $free;
 
         return array('total' => $total, 'free' => $free, 'used' => $used);
